Use different agama profiles for intel and arm builds
Primary goal is to enforce desired partitioning (avoid creating bios_boot
partition for intel builds + adjust size for EFI partition on arm).

+++ opensuse-leap16/opensuse-leap16.bootstrap.sh
@@ -70,7 +70,10 @@ passwd -d root
 sync
 
 # mumbo jumbo to give vmware-vdiskmanager more room for defragmenting and shrinking vmdk disk(s)
-for destpath in /boot /home/vagrant; do
+for destpath in /boot/efi /home/vagrant; do
+    if [ ! -d "${destpath}" ]; then
+        continue
+    fi
     touch ${destpath}/zeroes
     if ! dd if=/dev/zero of=${destpath}/zeroes bs=4k; then
         sync
